Changes to long

+++ src/SalesManager.java
@@ -1,13 +1,13 @@
 public class SalesManager {
-    protected int[] sales;
+    protected long[] sales;
 
-    public SalesManager(int[] sales) {
+    public SalesManager(long[] sales) {
         this.sales = sales;
     }
 
-    public int max() {
-        int max = -1;
-        for (int sale : sales) {
+    public long max() {
+        long max = -1;
+        for (long sale : sales) {
             if (sale > max) {
                 max = sale;
             }
@@ -15,9 +15,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = max();
-        for (int sale : sales) {
+    public long min() {
+        long min = max();
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -26,9 +26,9 @@ public class SalesManager {
     }
 
     public long statistics() {
-        int stat = 0;
-        int j = 0;
-        for (int sale : sales) {
+        long stat = 0;
+        long j = 0;
+        for (long sale : sales) {
             stat += sale;
                 j++;
             }
